Edit default message
This also changes the behavior so that if you do not have a custom file, it won't create one. Avoids bloat and avoids the situation we have to check for now, which is to have to hash the file to see if it's the same as the old SF message in order to replace it with this new message.

Also, the comment is indented now.